requirements.txt
streamlit
google-genai
textblob
pandas
python-dotenv
